Overhaul of our CI infrastructure: Moves all build jobs to a single workflow so we can properly declare dependencies between them.
Also removes the Generate Manifests action which adds a commit to all PRs. This is replaced by a job which regenerates charts and manifests and then fails the workflow if `git diff` shows any changes. This is based on the assumption that in that case obsolete crds, charts and/or manifests were committed.

Also adds various fixes:

- update helm & cargo deny actions
- remove dependabot and add renovate
- change bors to wait for an approving review
- remove (but not delete) helm roles.yaml so those can be managed in the operators
- fix shebang in generate manifests
- fix weird cp behavior for product config in makefile (subdirectory 'config-spec' should be removed now)
- add reviewdog workflow to run various linters on pull requests
- add udeps job to find unused dependencies