Hardcode the use of the "default" profile for rustup
At some point rustup changed the default profile from "default" to
"minimal" if no other profile was specified.

This caused our pre-commit checks to fail which rely on clippy which is
not in the "minimal" profile.
We now specify our dependency on the "default" profile explicitly to be
more independent of future rustup changes.